Refactor use admin rental hook at admin rental status
